Add files via upload
si

+++ JuanMendez_Lab7P2/src/juanmendez_lab7p2/Canciones.java
@@ -30,7 +30,7 @@ public class Canciones {
         this.titulo = titulo;
     }
 
-    public int getDuracion() {
+    public float getDuracion() {
         return duracion;
     }
